Allow to enable debug-mode (for sync) in options
In this mode, when a sync is run, ALL messages are outputed to the log.
So there is no real sync happening. There's also no need to have a
connected google account.

This is especially usefull for emulators testing new features with fake
sms and without google mail - to see what msg would get transferred.